commit 1
up to i was added to find random input from users both voice and user input..... till the code is working fine but if I give input capitalized it won't recognize. so planned to go for .. finding the keyword of that file name matching in all files it'll give result will print what are all the matching results found in all (C:/ , D:/ , E:/ ).